// src/interfaces/ProductDetailResponseDTO.ts
export interface ProductDetailResponseDTO {
    brandName: string;
    name: string;
    saledPrice: string;
    originalPrice: string;
    salePercent: string;
    mainImg: string;
    detailImg: string;
    detailImgHeight: number;
    detailImgWidth: number;
    isLiked: boolean;
    rate: number;
    reviewCount: string;
}